add docstrings for Proxy objects in datatypes/
